Introduce an optional default value in config.get()
Method config.get() function can now be called with a third (optional) value
if it's not Null and the option is not found, will be the returned value.

This allows the easy integration of new options which might be not (yet)
introduced in lime-defaults of a node after upgrading.

A Warning message is printed if default value used.

Signed-off-by: p4u <p4u@dabax.net>